deleted extra files, replaced some files
